+++ strcs/__init__.py
@@ -16,9 +16,9 @@ from .decorator import (
     WrappedCreator,
 )
 from .disassemble.base import Field, Type, TypeCache
-from .disassemble.hints import resolve_types
 from .disassemble.instance_check import InstanceCheck, InstanceCheckMeta
 from .disassemble.type_tree import MRO
+from .hints import resolve_types
 from .meta import Meta
 from .not_specified import NotSpecified, NotSpecifiedMeta
 from .register import CreateRegister, Creator, Registerer
