+++ strcs/hints.py
@@ -20,8 +20,8 @@ from attrs import fields as attrs_fields
 from attrs import has as is_attrs
 
 if tp.TYPE_CHECKING:
-    from ..register import CreateRegister
-    from .base import TypeCache
+    from .disassemble.base import TypeCache
+    from .register import CreateRegister
 
 T = tp.TypeVar("T")
 C = tp.TypeVar("C", bound=type)
@@ -256,7 +256,7 @@ def resolve_types(
 
     .. note:: Calling resolve_types will modify the fields on the class in place.
     """
-    from ..register import CreateRegister
+    from .register import CreateRegister
 
     if isinstance(type_cache, CreateRegister):
         type_cache = type_cache.type_cache
@@ -307,7 +307,7 @@ def resolve_types(
                     value = tp.ForwardRef(value, is_argument=False, is_class=True)
 
                 if name in allfields:
-                    from .base import Type
+                    from .disassemble.base import Type
 
                     disassembled = Type.create(value, cache=type_cache, expect=object)
